# mentorapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MentorAPI",
    "description": "REST API bridge for WinMentor DocImpServer COM interface v1.2.0\n\nAll endpoints except /api/health require API Key via header X-API-Key or query parameter api_key.\n\nMost data endpoints require a firma+luna to be selected first via POST /api/firme/select.\n\nAudit complet campuri: https://erp.malinco.ro/mentorapi/audit",
    "version": "1.3.0"
  },
  "servers": [
    {
      "url": "http://localhost:9500",
      "description": "Local"
    },
    {
      "url": "http://0.0.0.0:9500",
      "description": "Production"
    }
  ],
  "security": [
    {
      "ApiKeyHeader": []
    },
    {
      "ApiKeyQuery": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key"
      }
    }
  },
  "paths": {
...
    "/api/articole": {
      "get": {
        "tags": [
          "Articole"
        ],
        "summary": "Nomenclator articole (40 campuri/rand, paginat)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista articole paginata"
          }
        }
      }
    },
...
    "/api/stocuri": {
      "get": {
        "tags": [
          "Stocuri"
        ],
        "summary": "Stocuri articole (21 campuri/rand, paginat)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista stocuri"
          }
        }
      }
    },
...
  },
...
}